# Conflicts:
#	Cloudents.Core/Entities/ReadTutor.cs
#	Cloudents.Core/Entities/StudyRoomSession.cs
#	Cloudents.Persistance/Maps/ReadTutorMap.cs
#	Cloudents.Persistance/Maps/StudyRoomSessionMap.cs
#	Cloudents.Persistance/Repositories/ReadTutorRepository.cs
#	W-15-scripts.txt

+++ Cloudents.Core/Entities/CourseSubject.cs
@@ -1,6 +1,6 @@
 ﻿namespace Cloudents.Core.Entities
 {
-    public class CourseSubject : Entity<int>, IAggregateRoot
+    public class CourseSubject : Entity<long>, IAggregateRoot
     {
 
         protected CourseSubject()
